first hahn
1. moved declaration outside main

+++ process_sample.h
@@ -25,7 +25,7 @@ extern "C" {
     extern void processSamples(volatile unsigned int inputbuffer[2547], volatile unsigned char outputbuffer[2547][2], unsigned long pixelzahl)
     {
         volatile unsigned int inputbufferadress;
-        inputbufferadress =  &inputbuffer;  //Eingangsbufferadresse: 0x0856
+        inputbufferadress =  (int) &inputbuffer;  //Eingangsbufferadresse: 0x0856
         for (int highbyte = 0; highbyte<pixelzahl; highbyte++)
         {
            outputbuffer[highbyte][2]= (0x00FF && inputbuffer[highbyte]); //lowbyte extrahieren
